Mobility-Immobility Analysis
This script analyzes time-series speed data to identify periods of mobility and immobility
based on a speed threshold (0.02m/s). It processes the data to merge consecutive windows of the same
speed state that are shorter than a specified duration (less than 0.5s), calculates statistics for each window,
and visualizes the results.